feat(web): redesign the workspace-menu
